Sturdier solution to fix the peer's missing remoteId.
Handshake may come often or never, depending in which state the stream is passed in.
This means that .ready() should be triggered if a remoteId is available and only the first time that a handshake is submitted.
Multiple handshakes can technically update the remoteId. If the remoteId changes on update it is a good idea sending out a 'remote-update'
that triggers the change in remote identification.