Fix `BitInterlude` tests and edgecases (#735)
* Fix `BitInterlude` tests

Previously supported setting a default state for testing. This was
removed and the tests were not updated.

* Add failing test for edgecase

Currently a new range is inserted when setting a range that's not
contained in the current range, but overlaps and has a different value.
This insertion doesn't update the next range correctly so can report the
wrong value if the total ranges favor one range over another during
lookup.

The lookup is biased by making the bad range (aka the next range that's
not updated) the mid point during binary search.

* BitInterlude `setRange()` w/ new value updates both ranges when saddling

* Add test & fix for updating next range if different in `BitInterlude`