Add internal & external locks for bitfield operations
With the change to the bitfield being asynchronous, the previous
synchronous's implicit batch reading & writing ensure no conflicts
from writing while reading. To enforce that operations are sequential an
internal lock was added so if operations are called without awaiting or
are called between event loops because of an external message etc, they
will still not execute simultaneously.

To protect against read & write operations interleaving when they are
intended to be sequential, an external lock was added to claim the
bitfield roughly per protocol message. Because this theoretically will
cover the internal lock scenario by keeping access to a single chain of
async calls, it might be possible to remove the internal lock in the
future.

Ideally these wont be necessary but currently they solve the above
issues.