Lock on bitfield in `onsync` to ensure remote bitfield is updated
Without this lock, the `_update()` call in the `onsync` event doesn't
know about the remote bitfield update. So while `onsync` doesn't access
the `localBitfield` it does rely on updates to the `remoteBitfield`
which happen along with the `localBitfield` updates.

Fixes the 'restore after cancelled block request' test which could fail
because the sync after the new append wouldn't cause a `block` request
since the `b` peer assumed the `a` peer doesn't have the block. Since
the test waits for an append event (which doesn't mean the `upload`
event on `a` will be triggered) and the connection is destroyed
afterwards.