Add `.remoteFullyOpening` for guarding closing peer
Because `.broadcastRange()` called during `onopen` is now async, the
peer can be added to the replicator after a sync close is called on the
protomux channel. Since `onclose` assumed synchronous calls, it assumes
the peer is already added before it's closed. With it not added yet, the
peer isn't removed from the replicator and `replicator.destroy()` will
loop forever.

Added a destroy method to the bitfield for destroying the locks. Not
required for the fix, but should be reasonable to destroy
regardless.