Prevent undefined ranges reqs in `_updateNonPrimary()`
Now that checking the bitfield is async which makes iterating through
ranges async, the `_ranges` array can be mutated elsewhere while
awaiting. This means the current index of the range request can be
inaccurate when resolving the request.

To prevent this, the request index is looked up synchronously when the
request is resolved. This way the index is accurate and another request
isn't potentially clobbered by the popped head.

Since this logic already exists to unref the request (for gc'ing &
cancelling), the `_unref()` is reused. A success boolean is needed to
update the index in the `_updateNonPrimary()` ranges loop. So all
request `_unref()`s are updated to return a success bool.