Wait for localBitfield locks to flush before closing session/core
This prevents the `session id reuse does not stall` test from failing
randomly, usually with only `99` download events instead of the expected
`100`. This is because processing ondata requests requires more event
loop iterations to update / check the bitfield in storage. This allows
the close promise for a session to resolve while the request is
processed meaning the test checks the number of downloads before all
events can be emitted.