Update remoteContiguousLength only on range starting at `0`
Peers always send `range` requests starting with `0` if they are
contiguous. Partial ranges would imply they are not contiguous so
shouldn't be able to sparsely update the `remoteContiguousLength`.

Because there is no need check the `start` of the range &
`core.updateRemoteContiguousLength()` already checks if the `length` is
larger, the logic is simplified to only call
`updateRemoteContiguousLength()` when updating the
`_remoteContiguousLength` for the peer when the range isnt a drop.